Implemented time Engine, and provided a means to use it via GameEngine. Removed the need for "out" on all time interfaces, and updated body parts accordingly.  Moved all time related stuff to new Time subfolder in the Engine folder, updated namespaces accordingly. Cleaned up all code that errored in backend,

+++ CoC.Backend/Attacks/BodyPartAttacks/GenericBite.cs
@@ -12,7 +12,7 @@ namespace CoC.Backend.Attacks.BodyPartAttacks
 		private readonly byte biteStrength;
 		private readonly SimpleDescriptor teeth;
 
-		protected override ushort attackCost => ((byte)biteStrength.mult(5)).add(10);
+		protected override ushort attackCost => biteStrength.mult(5).add(10);
 
 		public GenericBite(SimpleDescriptor teethName, byte strength = 1) : base(Attack)
 		{
